requirements.txt
# LangChain相关依赖
langchain>=0.3.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
tavily-python>=0.2.4
langchain-tavily>=0.1.0

# 向量存储相关依赖
pymilvus>=2.0.0

# 文档处理相关依赖
unstructured>=0.10.0
PyPDF2>=3.0.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
markdown>=3.4.0

# FastAPI相关依赖
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5
